mi_app/cloud_run/app/requirements.txt
dash
dash-dcc
dash-html-components
dash-table
google-cloud-firestore
google-cloud-storage